cut functions file into required files
